refactoring
Some refactoring, some filenames have been changed, comments added,
small changes in the classes.
