--- ProjectCource/nbproject/private/private.properties
@@ -1,6 +1,0 @@
-auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
-# No need to modify this property unless customizing JavaFX Ant task infrastructure
-endorsed.javafx.ant.classpath=.
-javafx.run.inbrowser=<Default System Browser>
-javafx.run.inbrowser.path=/usr/bin/open
-user.properties.file=/Users/beysimeryalmaz/Library/Application Support/NetBeans/8.2/build.properties
