--- ProjectCource/nbproject/private/private.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/View/MainView.fxml</file>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/View/ViewFactory.java</file>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/View/addTimelineView.fxml</file>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/ProjectCource.java</file>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/Controller/MainWindowController.java</file>
-            <file>file:/Users/beysimeryalmaz/NetBeansProjects/ProjectCource/src/projectcource/Controller/addTimelineController.java</file>
-        </group>
-    </open-files>
-</project-private>
