--- ProjectCource/src/projectcource/Controller/addTimelineController.java
@@ -1,35 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package projectcource.Controller;
-
-import com.jfoenix.controls.JFXButton;
-import com.jfoenix.controls.JFXDatePicker;
-import com.jfoenix.controls.JFXTextArea;
-import com.jfoenix.controls.JFXTextField;
-import javafx.fxml.FXML;
-
-/**
- *
- * @author beysimeryalmaz
- */
-public class addTimelineController {
-    
-    @FXML
-    private JFXTextField titleField;
-
-    @FXML
-    private JFXDatePicker startDate;
-
-    @FXML
-    private JFXDatePicker endDate;
-
-    @FXML
-    private JFXTextArea descriptionField;
-
-    @FXML
-    private JFXButton saveButton;
-    
-}
