+++ ProjectFiles/.idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,5 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+  </profile>
+</component>
