+++ ProjectFiles/src/timelineManager/controller/AddTimelineController.java
@@ -1,0 +1,31 @@
+package timelineManager.controller;
+
+import com.jfoenix.controls.JFXButton;
+import com.jfoenix.controls.JFXDatePicker;
+import com.jfoenix.controls.JFXTextArea;
+import com.jfoenix.controls.JFXTextField;
+import javafx.fxml.FXML;
+
+/**
+ * This is a controller class for the AddTimelineView.
+ * it holds the variables and logics needed to create a new timeline from GUI.
+ * @author beysimeryalmaz
+ */
+public class AddTimelineController {
+    
+    @FXML
+    private JFXTextField titleField;
+
+    @FXML
+    private JFXDatePicker startDate;
+
+    @FXML
+    private JFXDatePicker endDate;
+
+    @FXML
+    private JFXTextArea descriptionField;
+
+    @FXML
+    private JFXButton saveButton;
+    
+}
