+++ ProjectFiles/src/timelineManager/model/TimelineModel.java
@@ -1,0 +1,26 @@
+package timelineManager.model;
+
+import java.util.ArrayList;
+
+/**
+ *  This class holds an ArrayList of Timelines
+ */
+public class TimelineModel
+{
+    ArrayList<Timeline> timelines;
+    
+    public ArrayList<Timeline> getTimelines()
+    {
+        return timelines;
+    }
+    
+    public void setTimelines(ArrayList<Timeline> timelines)
+    {
+        this.timelines = timelines;
+    }
+    
+    TimelineModel() {}
+    
+   
+    
+}
