+++ ProjectFiles/src/timelineManager/view/AddTimelineView.fxml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import com.jfoenix.controls.JFXDatePicker?>
+<?import com.jfoenix.controls.JFXTextArea?>
+<?import com.jfoenix.controls.JFXTextField?>
+<?import javafx.scene.layout.AnchorPane?>
+
+
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="350.0" style="-fx-background-color: b2c7e8;" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.111" fx:controller="timelineManager.controller.AddTimelineController">
+   <children>
+      <JFXTextField fx:id="titleField" labelFloat="true" layoutX="37.0" layoutY="22.0" prefHeight="29.0" prefWidth="284.0" promptText="Title" />
+      <JFXDatePicker fx:id="startDate" editable="false" layoutX="37.0" layoutY="250.0" prefHeight="29.0" prefWidth="292.0" promptText="Start Date" />
+      <JFXDatePicker fx:id="endDate" editable="false" layoutX="37.0" layoutY="288.0" prefHeight="29.0" prefWidth="292.0" promptText="End Date" />
+      <JFXTextArea fx:id="descriptionField" labelFloat="true" layoutX="37.0" layoutY="86.0" prefHeight="126.0" prefWidth="284.0" promptText="Description" />
+      <JFXButton fx:id="saveButton" buttonType="RAISED" layoutX="103.0" layoutY="359.0" prefHeight="27.0" prefWidth="144.0" style="-fx-background-color: edc6bd;" text="Save" />
+   </children>
+</AnchorPane>
